fix pre tag bg color in markdown (#678)
