+++ src/app/common/elements/markdown.less
@@ -46,7 +46,7 @@
         padding: 2px 4px 2px 6px;
     }
 
-    pre.codeblock {
+    pre {
         background-color: var(--markdown-bg-color);
         margin: 4px 10px;
         padding: 0.4em 0.7em;
